backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlmodel==0.0.22
psycopg2-binary==2.9.9
alembic==1.13.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File upload & env
python-multipart==0.0.9
python-dotenv==1.0.1

# LLM / RAG
openai==1.51.2
chromadb==0.5.5
pymupdf==1.24.10

# HTTP
httpx==0.27.0
